% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioage.R
\name{gold_bioage}
\alias{gold_bioage}
\title{gold_bioage}
\usage{
gold_bioage(d4, var)
}
\arguments{
\item{d4}{Dataframe,containing survival information (time, status)}

\item{var}{Variables for developing biological age (Age and Biomarkers)}
}
\value{
a list
\itemize{
  \item coef1, Model 1 cofficients
  \item coef2, Model 2 cofficients
  \item coefficients, Biological age cofficients
  \item Bioage, Biological age data.frame
}
}
\description{
Develop biological age through Gompertz hazard function
\itemize{
  \item Model 1: Gompertz regression with Surv(time, status) ~ age
  \item Model 2: Gompertz regression with Surv(time, status) ~ age + Biomarkers
  }
}
\examples{
head(NHANES4)
var <- c(
  "age", "albumin", "alp", "creat",
  "glucose_mmol", "lymph",
  "mcv", "rdw", "wbc", "ggt"
)

bioage <- gold_bioage(NHANES4, var)
}
