% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHANES4}
\alias{NHANES4}
\title{NHANES4}
\format{
An object of class \code{data.frame} with 54945 rows and 30 columns.
}
\usage{
NHANES4
}
\description{
NHANES4 1998-2018 biomakers data
}
\examples{
head(NHANES4)
}
\keyword{datasets}
